requirements.txt
absl-py==0.10.0
altgraph==0.17
AnyQt==0.0.11
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828585465/work
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
auto-py-to-exe==2.7.8
autorank==1.0.1
backcall==0.2.0
baycomp==1.0.2
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
bottle==0.12.18
bottle-websocket==0.2.9
Bottleneck==1.3.2
brminer==0.0.3
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1600699246375/work
chardet==3.0.4
click==7.1.2
colorama==0.4.3
combo==0.1.1
commonmark==0.9.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dictdiffer==0.8.1
docutils==0.16
Eel==0.12.4
emoji==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
future==0.18.2
gast==0.3.3
gevent==20.6.2
gevent-websocket==0.10.1
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.32.0
h11==0.9.0
h5py==2.10.0
httpcore==0.10.2
httpx==0.14.3
idna==2.10
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1598865816163/work
ipython-genutils==0.2.0
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
isort==5.4.2
jdcal==1.4.1
jedi @ file:///C:/ci/jedi_1598353661586/work
Jinja2==2.11.2
joblib==0.16.0
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==21.4.0
keyrings.alt==4.0.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.34.0
lockfile==0.12.2
Markdown==3.3.1
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
msgpack==1.0.0
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601914925608/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1601499549014/work
networkx==2.5
nltk==3.5
notebook @ file:///C:/ci/notebook_1601501643625/work
numba==0.51.2
numpy==1.18.5
oauthlib==3.1.0
openpyxl==3.0.5
openTSNE==0.4.4
opt-einsum==3.3.0
orange-canvas-core==0.1.18
orange-widget-base==4.9.0
Orange3==3.27.0
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pylint==2.6.0
pyod==0.8.3
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.1
pyqtgraph==0.11.0
PyQtWebEngine==5.15.1
pyrsistent @ file:///C:/ci/pyrsistent_1600141795814/work
PySocks==1.7.1
python-dateutil==2.8.1
python-louvain==0.14
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.4.0
rsa==4.6
scikit-learn==0.23.2
scikit-posthocs==0.6.6
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
serverfiles==0.3.0
sip==4.19.13
six==1.15.0
sklearn==0.0
sniffio==1.2.0
statsmodels==0.12.0
suod==0.0.4
sympy==1.6.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.48.2
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
tweepy==3.9.0
Unidecode==1.1.1
urllib3==1.25.10
vaderSentiment==3.3.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wincertstore==0.2
wordcloud==1.8.0
wrapt==1.12.1
xgboost==1.2.1
xlrd==1.2.0
XlsxWriter==1.3.7
zipp @ file:///tmp/build/80754af9/zipp_1602002379921/work
zope.event==4.4
zope.interface==5.1.0
